refactor(auth): Refresh Token JPA 구현체를 domain 계층으로 이동
infrastructure -> model 직접 의존을 제거하고, JPA 기반 구현은 D -> M 만 허용된
계층 규칙에 맞게 domain/auth 로 옮긴다. JWT/Caffeine 같은 외부 라이브러리
어댑터만 infrastructure 에 남긴다.
